font: render preview glyphs for non-latin fonts too
We try to be smart here, and if we find out the font doesn't have glyphs
to represent the characters we want to preview, we build a map of all
the glyphs in the font, choose a random set and display those.

The initial lowercase/uppercase/punctuation alphabets are omitted entirely
in case there's no glyphs available to render them.