media-bin: destroy video widget on sink deinit
Right now, when the video sink is deinitialized, our reference to the
video widget is dropped by unreffing it.
There's an expectation that the resources used by the widget will be
released after that, but that may not actually happen as something
else could still hold another reference to the object.

This commit always calls gtk_widget_destroy() on the video widget
on deinit, to make sure nothing is left behind.

This fixes fullscreening the video player under Wayland.