Port to use package.js
Let's finally remove the C wrapper and port the app to use the
package.js framework and simplify greatly the way the app is installed
and runs.

As part of this, we also move the JS code to a gresource and change
some of the installation paths to account for the structure required
by package.js.
In practice, this means changing the package name to
org.gnome.NautilusPreviewer, but we keep the original sushi binary
name installed in ${bindir} for compatibility.