evince: load EvinceDocument from JS
Instead of having SushiPdfLoader, a class to load the EvinceDocument
and do conversion from LibreOffice when required, implement loading
natively in the viewer, and factor out an async method for the LO
conversion.

The goal here is to simplify the viewer state and slim down libsushi
as much as possible.