text: install and use builder-dark style
Instead of reading and using the gedit scheme. This is necessary
since we use the dark theme, and gedit may not, leading to unreadable
text.

In the future we may tweak this to have two styles (one for light
and one for dark variants), or something else, but in the meantime
this provides a big improvement in legibility.

https://gitlab.gnome.org/GNOME/sushi/issues/24