audio: Fetch ASIN with libsoup instead of libmusicbrainz
We are only using libmusicbrainz for finding the ASIN code for a
given artist and album.

With this changes we use libsoup to query the musicbrainz API
directly, allowing for the dependency on libmusicbrainz5 to be
dropped.

See https://musicbrainz.org/doc/MusicBrainz_API
and https://musicbrainz.org/doc/ASIN