sushi-media-bin: Fix position update on fullscreen
We have to add the tick callback to the overlay widget instead of the main
widget, otherwise the position bar is not updated in fullscreen mode when the
gtksink is used. This happens because we move the overlay between the main and
fullscreen windows, and we did't get the events when the overlay is outside of
the main widget.